Made some minor changes to modules
