+++ src/modules/globals.ts
